Tentative fix for KNTK-306 issue (#58)
Fix for KNTK-306

The root cause of the problem is duplication in `x` and `y` labels of the matrix which is in turn caused by duplicate agent names. Plotly ignores rows and colums with duplicate labels. The fix uses agent names provided in the `mesh` section of augmented test health results which are more usable (mostly unique) than agent names provide by the 'AgentsList' API operation. This also makes agents names identical to those shown in the matrix display of the same mesh test in the Kentik UI. The fix also checks uniqueness of agent names and if a duplicate is detected appends agent id to names of both colliding agents.

The inconsistency of agent names in Kentik API will (hopefully) be fixed soon, at which point  this workaround will become unnecessary, but (should) continue to work.

Contains also:
- refactoring of the Agents cache (sorted iteration over agents now sorts by agent name)
- correction of logical error in SyntheticsRepo._get_mesh_test_rows
- type annotation improvements
- additional warning and debug log messages